requirements.txt
robotframework==2.9.2
paramiko==1.16.0
scp==0.10.2
ipaddress==1.0.16
interruptingcow==0.6
PyYAML==3.11
pykwalify==1.5.0
scapy==2.3.1

